Update read and write bytes functions to takes uint64 instead of int

+++ kernel/functions.h
@@ -7,10 +7,10 @@ void            brelse(Buffer *);
 void            bwrite(Buffer *);
 void            bpin(Buffer *);
 void            bunpin(Buffer *);
-void            readBytes(int, int, int, uchar *);
-void            writeBytes(int, int, int, uchar *);
-void            readBytesVirtual(int, int, int, uchar *, PageTable *);
-void            writeBytesVirtual(int, int, int, uchar *, PageTable *);
+void            readBytes(uint64, uint64, uint64, uchar *);
+void            writeBytes(uint64, uint64, uint64, uchar *);
+void            readBytesVirtual(uint64, uint64, uint64, uchar *, PageTable *);
+void            writeBytesVirtual(uint64, uint64, uint64, uchar *, PageTable *);
 
 // console.c
 void            consoleinit(void);
